Move Pagination to own folder.
Fix WithRequest reloading bug.
