+++ src/elements/Pagination/style.less
@@ -1,0 +1,75 @@
+.pagination {
+  display: flex;
+  padding: 2rem 0;
+  justify-content: space-between;
+  align-items: center;
+
+  @media @phone {
+    flex-wrap: wrap;
+  }
+  a, > span {
+    color: black;
+
+    &:hover {
+      color: currentColor;
+      opacity: 0.8;
+    }
+
+    &.disabled {
+      opacity: 0.4;
+
+      &:hover {
+        opacity: 0.4;
+      }
+    }
+
+    span.arrow {
+      display: inline-block;
+      border: solid currentColor;
+      border-width: 0 2px 2px 0;
+      padding: 4px;
+
+      &.next {
+        transform: rotate(-45deg);
+        margin-right: 4px;
+      }
+
+      &.back {
+        transform: rotate(135deg);
+        margin-left: 4px;
+      }
+    }
+  }
+
+  ul {
+    list-style: none;
+    display: block;
+    padding: 0;
+    margin: 0;
+
+    @media @phone {
+      order: 3;
+      width: 100%;
+      text-align: center;
+      padding-top: 1.5rem;
+    }
+
+    li {
+      display: inline;
+      padding: 0 0.5rem;
+      margin-bottom: 0;
+
+      a {
+        color: @light-grey;
+
+        &.active {
+          color: @white;
+          font-weight: 500;
+          background-color: @primary-color;
+          padding: 0.2rem 0.65rem;
+          border-radius: 4px;
+        }
+      }
+    }
+  }
+}
