+++ src/hoc/WithRequest.js
@@ -66,15 +66,21 @@ export default class WithRequest extends React.Component {
   }
 
   shouldComponentUpdate (nextProps, nextState) {
-    const nextEnpoint = (nextProps.request || {}).endpoint
-    const currEnpoint = (this.props.request || {}).endpoint
-    if (currEnpoint !== nextEnpoint) {
+    const nextEndpoint = (nextProps.request || {}).endpoint
+    const currEndpoint = (this.props.request || {}).endpoint
+    if (currEndpoint !== nextEndpoint) {
       return true
     }
     return !equal(nextState, this.state)
   }
 
   componentDidUpdate (prevProps, prevState, snapshot) {
+    const endpointChanged = this.props.request !== prevProps.request
+    if (endpointChanged) {
+      this.setState({isLoading: true, result: null})
+      this._loadResult(this.props)
+      return
+    }
     if (!this._existing) return
     if ((this.props.request || {}).endpoint !== this._existing._endpoint) {
       this.setState({isLoading: true})
